chore: refactor the source module into a new manager module
